update api gateway 1.1.2

+++ api/v2/service/validators.go
@@ -10,7 +10,7 @@ import (
 )
 
 func (s *Service) Validators(_ context.Context, req *pb.ValidatorsRequest) (*pb.ValidatorsResponse, error) {
-	height := req.Height
+	height := int64(req.Height)
 	if height == 0 {
 		height = int64(s.blockchain.Height())
 	}
